requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.8.1
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0